@@ -18,12 +18,19 @@
   "$username": "onface",
   "$repository": "radio.react",
   "dependencies": {
-    "checkbox.react": "^0.3.10",
-    "extend": "^3.0.1",
+    "checkbox.react": "^1.0.0",
+    "classnames": "^2.2.5",
+    "safe-extend": "^3.0.5",
+    "util.react": "^1.6.0",
     "prop-types": "^15.6.0",
-    "util.react": "^1.4.1"
+    "react-spread-props": "^0.1.1"
+  },
+  "optionalDependencies": {
+    "react": "^16.4.0",
+    "react-dom": "^16.4.0"
   },
   "devDependencies": {
+    "codesandbox": "^1.2.10",
     "babel": "^6.23.0",
     "babel-core": "^6.26.0",
     "babel-loader": "^7.1.2",
@@ -38,13 +45,8 @@
     "babel-polyfill": "^6.26.0",
     "babel-preset-es2015": "^6.24.1",
     "chai": "^4.1.2",
-    "css-loader": "^0.26.1",
-    "ejs": "^2.5.7",
-    "enzyme": "^3.3.0",
-    "enzyme-adapter-react-16": "^1.1.1",
     "express": "^4.16.2",
     "extend": "^3.0.1",
-    "file-loader": "^1.1.6",
     "fis-parser-less-2.x": "^0.1.4",
     "fis3-hook-relative": "^2.0.1",
     "fis3-parser-vue-component": "^5.5.1",
@@ -53,7 +55,19 @@
     "hash-to-port": "^1.0.0",
     "is-absolute-url": "^2.1.0",
     "jasmine-core": "^2.8.0",
+    "less-plugin-autoprefix": "^1.5.1",
+    "less-plugin-functions": "^1.0.0",
+    "livereload": "^0.6.3",
+    "markrun": "^0.23.0",
+    "open": "^0.0.5",
+    "vue-template-compiler": "^2.5.13",
+    "css-loader": "^0.26.1",
+    "file-loader": "^1.1.6",
     "json-loader": "^0.5.7",
+    "url-loader": "^0.6.2",
+    "ejs": "^2.5.7",
+    "enzyme": "^3.3.0",
+    "enzyme-adapter-react-16": "^1.1.1",
     "json5": "^0.5.1",
     "karma": "^2.0.0",
     "karma-babel-preprocessor": "^7.0.0",
@@ -64,21 +78,9 @@
     "karma-webpack": "^2.0.9",
     "less": "^2.7.3",
     "less-loader": "^2.2.3",
-    "less-plugin-autoprefix": "^1.5.1",
-    "less-plugin-functions": "^1.0.0",
-    "livereload": "^0.6.3",
-    "markrun": "^0.23.0",
-    "open": "^0.0.5",
-    "prop-types": "^15.6.0",
-    "react": "^16.2.0",
-    "react-dom": "^16.2.0",
     "react-hot-loader": "^4.0.0-beta.6",
     "style-loader": "^0.13.1",
-    "url-loader": "^0.6.2",
-    "util.react": "^1.4.1",
-    "vue": "^2.5.13",
     "vue-loader": "^13.6.1",
-    "vue-template-compiler": "^2.5.13",
     "webpack": "^3.10.0",
     "webpack-dev-middleware": "^2.0.3",
     "webpack-hot-middleware": "^2.21.0"
