@@ -14,17 +14,20 @@ class RadioGroup extends Component {
             util.themes(props)
         ].join(' ')
 
-        let children = extend(true,[],self.props.children)
-        children.map(function (item, index) {
-            item.props.name = item.props.name || self.props.name
-            item.key = item.key || index
-            item.props.checked = item.props.value === self.props.value
-            item.props.onChange = () => {
-                self.props.onChange(item.props.value)
-            }
-            return item
+        // let children = extend(true,[],self.props.children)
+        // children.map(function (item, index) {
+        //     item.props.name = item.props.name || self.props.name
+        //     item.key = item.key || index
+        //     item.props.checked = item.props.value === self.props.value
+        //     item.props.onChange = () => {
+        //         self.props.onChange(item.props.value)
+        //     }
+        //     return item
+        // })
+        console.log(React.children)
+        React.Children.map(function(item,index){
+            console.log(index,':',item)
         })
-
         return (
             <div
                 className={cls({
@@ -35,7 +38,11 @@ class RadioGroup extends Component {
                     [`${pcls}--hideicon`] : props.hideicon ,
                 })}
             >
-                {children}
+                {
+                    self.props.render
+                    ? self.props.render(self)
+                    : self.props.children
+                }
             </div>
 
         )
